misc/HtmlTestSuite/results/test-0026.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>Largest Contentful Paint: observe image inside SVG</title>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="resources/largest-contentful-paint-helpers.js"/>
<script>
async_test(function (t) {
  assert_implements(window.LargestContentfulPaint, "LargestContentfulPaint is not implemented");
  const beforeLoad = performance.now();
  new PerformanceObserver(
    t.step_func_done(entryList =&gt; {
      assert_equals(entryList.getEntries().length, 1);
      const entry = entryList.getEntries()[0];
      const url = window.location.origin + '/images/blue.png';
      // blue.png is 133 by 106.
      const size = 133 * 106;
      checkImage(entry, url, 'image_id', size, beforeLoad);
    })
  ).observe({type: 'largest-contentful-paint', buffered: true});
}, "Image inside SVG is observable.");
</script>
</head><body><svg xmlns="http://www.w3.org/2000/svg" height="300" id="svg_id" width="300">
  <image href="/images/blue.png" id="image_id"/>
</svg>
</body></html>
